Move some data into battle_bg.c
