+++ ld_script.txt
@@ -493,7 +493,8 @@ SECTIONS {
         src/rtc.o(.rodata);
         src/main_menu.o(.rodata);
         data/mon_attrs.o(.rodata);
-	src/battle/battle_1.o(.rodata);
+        src/battle/battle_1.o(.rodata);
+        src/battle/battle_bg.o(.rodata);
         data/btl_attrs.o(.rodata);
         src/battle/battle_4.o(.rodata);
         src/battle/battle_controller_player.o(.rodata);
